Fix for div,span,td mask function
Fix the value used to evaluate mask if mask is a function and the element is not an input (div, td, span)

The el.val() returns a value only if it's an input. In case of (i.e.) span it return a value of "" and the function doesn't receive the correct value.