Fix onChange not firing when typing previous value
The current implementation is not firing the onChange callback passed in
the options.

This happens because the oldValue variable is updated only on blur. With
this changes, two variables are used. One controls the value that the
element had when it got focus, and the other one is changed immediately
after every change.